prebunk/requirements.txt
firebase_admin
djangorestframework
markdown
django-filter
python-dotenv
django-cors-headers
gunicorn
